change object structure of gatlinggun
 GatlingGun is now a module rather than a class. The module can now take
a configure block. Further, what was in the original GatlingGun class is
now in GatlingGun::Client. GatlingGun.client can be accessed after
configuring the module to access a preconfigured GatlingGun::Client
